+++ src/main/java/org/apache/maven/plugins/enforcer/RestrictImports.java
@@ -40,6 +40,7 @@ public class RestrictImports extends BannedImportGroupDefinition implements Enfo
 
     private static final String SKIP_PROPERTY_NAME = "restrictimports.skip";
     private static final String FAIL_BUILD_PROPERTY_NAME = "restrictimports.failBuild";
+    private static final String PARALLEL_ANALYSIS_PROPERTY_NAME = "restrictimports.parallel";
 
     private final SourceTreeAnalyzer analyzer = SourceTreeAnalyzer.getInstance();
     private final MatchFormatter matchFormatter = MatchFormatter.getInstance();
@@ -52,6 +53,7 @@ public class RestrictImports extends BannedImportGroupDefinition implements Enfo
     private List<File> excludedSourceRoots = new ArrayList<>();
     private boolean failBuild = true;
     private boolean skip = false;
+    private boolean parallel = false;
 
     @Override
     public EnforcerLevel getLevel() {
@@ -75,7 +77,7 @@ public class RestrictImports extends BannedImportGroupDefinition implements Enfo
             final BannedImportGroups groups = createGroupsFromPluginConfiguration();
             LOGGER.debug("Banned import groups:\n{}", groups);
 
-            final AnalyzerSettings analyzerSettings = createAnalyzerSettingsFromPluginConfiguration(project);
+            final AnalyzerSettings analyzerSettings = createAnalyzerSettingsFromPluginConfiguration(helper, project);
             LOGGER.debug("Analyzer settings:\n{}", analyzerSettings);
 
             final AnalyzeResult analyzeResult = analyzer.analyze(analyzerSettings, groups);
@@ -123,7 +125,8 @@ public class RestrictImports extends BannedImportGroupDefinition implements Enfo
     }
 
     private AnalyzerSettings createAnalyzerSettingsFromPluginConfiguration(
-            MavenProject mavenProject) {
+            EnforcerRuleHelper helper,
+            MavenProject mavenProject) throws Exception {
         if (!(includeCompileCode || includeTestCode)) {
             throw new IllegalArgumentException("Configuration error: No sources were included");
         }
@@ -137,11 +140,13 @@ public class RestrictImports extends BannedImportGroupDefinition implements Enfo
                 : Collections.emptyList();
 
         final Charset sourceFileCharset = determineSourceFileCharset(mavenProject);
+        final boolean parallel = isParallel(helper);
 
         return AnalyzerSettings.builder()
                 .withSrcDirectories(srcDirectories)
                 .withTestDirectories(testDirectories)
                 .withSourceFileCharset(sourceFileCharset)
+                .enableParallelAnalysis(parallel)
                 .build();
     }
 
@@ -300,6 +305,21 @@ public class RestrictImports extends BannedImportGroupDefinition implements Enfo
             return "true".equalsIgnoreCase(skipProperty.toString());
         }
         return this.skip;
+    }
+
+    public void setParallel(boolean parallel) {
+        this.parallel = parallel;
+    }
+
+    private boolean isParallel(EnforcerRuleHelper evaluator) throws Exception {
+        final Object parallelProperty = evaluator.evaluate("${" + PARALLEL_ANALYSIS_PROPERTY_NAME + "}");
+        if (parallelProperty != null) {
+            LOGGER.warn(
+                    "'{}={}' has been passed which takes precedence over 'parallel={}' configuration in the pom file",
+                    PARALLEL_ANALYSIS_PROPERTY_NAME, parallelProperty, this.parallel);
+            return "true".equalsIgnoreCase(parallelProperty.toString());
+        }
+        return this.parallel;
     }
 
     @Override